refactor auth handler(s)
